+++ src/js/main/constants/index.js
@@ -1,0 +1,12 @@
+// Scene: Setup
+export const chunkSize = 32;
+export const worldSize = 40;
+export const chunkSliceSize = chunkSize * chunkSize;
+export const fov = 75;
+export const aspect = 2; // the canvas default
+export const near = 0.1;
+export const far = 1000;
+
+export const tileSize = 16;
+export const tileTextureWidth = 256;
+export const tileTextureHeight = 64;
