+++ src/js/main/controls/currentObjectId.js
@@ -1,0 +1,3 @@
+let currentObjectId;
+
+export default currentObjectId;
