+++ src/js/main/controls/speedSlider.js
@@ -1,0 +1,3 @@
+const speedSlider = document.getElementById("speedslider");
+
+export default speedSlider;
