+++ src/js/main/controls/worlds.js
@@ -1,0 +1,3 @@
+const worlds = {};
+
+export default worlds;
